requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==19.9.0
matplotlib==3.1.1
numpy==1.17.2
Pillow==6.2.0
http://download.pytorch.org/whl/cpu/torch-0.4.1-cp36-cp36m-linux_x86_64.whl
torchvision==0.4.0